Seperated health and mana regen for non-mana champ scalability.
