+++ Assets/Scripts/HealthRegen.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 24f1bba863ab4cf438e0ced9e40efa4e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
