+++ Assets/Scripts/ManaRegen.cs
@@ -2,12 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-/*
-* Purpose: Handles health and mana regeneration.
-*
-* @author: Colin Keys
-*/
-public class HealthManaRegen : MonoBehaviour
+public class ResourceRegen : MonoBehaviour
 {
     private Player player;
     private ChampionStats championStats;
@@ -22,22 +17,16 @@ public class HealthManaRegen : MonoBehaviour
     {
         championStats = (ChampionStats) player.unitStats;
         StartCoroutine(RegenMana());
-        StartCoroutine(RegenHealth());
     }
 
+    // Called once per frame.
     void Update(){
-        // Activate mana and health regen UI if the player is not full mana or full health.
+        // Activate mana regen UI if the player is not full mana.
         if(championStats.currentMana < championStats.maxMana.GetValue() && !player.isDead){
             UIManager.instance.SetManaRegenActive(true, player.playerUI);
         }
         else{
             UIManager.instance.SetManaRegenActive(false, player.playerUI);
-        }
-        if(championStats.currentHealth < championStats.maxHealth.GetValue() && !player.isDead){
-            UIManager.instance.SetHealthRegenActive(true, player.playerUI);
-        }
-        else{
-            UIManager.instance.SetHealthRegenActive(false, player.playerUI);
         }
     }
 
@@ -63,33 +52,6 @@ public class HealthManaRegen : MonoBehaviour
                 //uiManager.UpdateManaBar();
                 // Display per 1s on UI.
                 UIManager.instance.UpdateManaRegen(championStats.MP5.GetValue()/5.0f, player.playerUI);
-            }
-            yield return new WaitForSeconds(0.5f);
-        }
-    }
-
-    /*
-    *   RegenHealth - Regens the champions health every 0.5s based on their HP5 stat.
-    */
-    private IEnumerator RegenHealth(){
-        float healthToRegen;
-        while(gameObject){
-            // 0.5s intervals over 5s = 10 increments.
-            healthToRegen = Mathf.Round(championStats.HP5.GetValue()/10.0f);
-            float maxHealth = championStats.maxHealth.GetValue();
-            float currentHealth = championStats.currentHealth;
-            // If not at max mana and not dead then regen health.
-            if(currentHealth < maxHealth && !player.isDead){
-                currentHealth += healthToRegen;
-                // If regen goes over max health set current health to max.
-                if(currentHealth > maxHealth)
-                    currentHealth = maxHealth;
-                // Set health and only regen every 0.5s.
-                championStats.SetHealth(currentHealth);
-                UIManager.instance.UpdateHealthBar(player, player.playerUI, player.playerBar);
-                //uiManager.UpdateHealthBar();
-                // Display per 1s on UI.
-                UIManager.instance.UpdateHealthRegen(championStats.HP5.GetValue()/5.0f, player.playerUI);
             }
             yield return new WaitForSeconds(0.5f);
         }
