+++ Assets/Scripts/ManaRegen.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 79736f83d4f6713408686f464edb94ce
+guid: 23f47262ce278cc4696868123cf38913
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
